+++ src/main/java/com/github/thedeathlycow/moregeodes/MoreGeodesClient.java
@@ -1,9 +1,6 @@
 package com.github.thedeathlycow.moregeodes;
 
-import com.github.thedeathlycow.moregeodes.blocks.Cutouts;
-import com.github.thedeathlycow.moregeodes.blocks.ModBlocks;
-import com.github.thedeathlycow.moregeodes.features.ModFeatures;
-import com.github.thedeathlycow.moregeodes.item.ModItems;
+import com.github.thedeathlycow.moregeodes.client.GeodesCutouts;
 import net.fabricmc.api.ClientModInitializer;
 
 public class MoreGeodesClient implements ClientModInitializer {
@@ -11,7 +8,7 @@ public class MoreGeodesClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         MoreGeodes.LOGGER.info("Beginning More Geodes client initialization");
-        Cutouts.registerCutouts();
+        GeodesCutouts.registerCutouts();
         MoreGeodes.LOGGER.info("More Geodes client initialized!");
     }
 }
